+++ libs/WickrIOLib/createjson.h
@@ -8,8 +8,8 @@ class DECLSPEC CreateJsonAction
 {
 public:
     CreateJsonAction();
-    CreateJsonAction(QString action, QStringList users, int ttl, QString message, QList<QString> attachment);
-    CreateJsonAction(QString action, QString name, int ttl, QString message, QList<QString> attachment, bool isVGroupID = false);
+    CreateJsonAction(QString action, QStringList users, long ttl, QString message, QList<QString> attachment);
+    CreateJsonAction(QString action, QString name, long ttl, QString message, QList<QString> attachment, bool isVGroupID = false);
     ~CreateJsonAction();
 
     void setBOR(int bor);
@@ -19,8 +19,8 @@ public:
     QStringList users;
     QString name;
     QString vgroupid;
-    int ttl;
-    int m_bor;
+    long ttl;
+    long m_bor;
     bool m_has_bor;
     QString message;
     QList<QString> attachments;
